requirements.txt
PyYAML==5.3
requests==2.24.0
Click==7.0
validators==0.18.1
